# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "My API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "請輸入 Bearer <token>"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/signin": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "用戶登入",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "使用者登入資料",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string",
                  "example": "johndoe"
                },
                "password": {
                  "type": "string",
                  "example": "123456"
                }
              },
              "required": [
                "username",
                "password"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "登入成功",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "array",
                  "example": [],
                  "items": {}
                },
                "token": {
                  "type": "string",
                  "example": "jwt_token"
                },
                "message": {
                  "type": "string",
                  "example": "Authentication successed"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "登入失敗",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "array",
                  "example": [],
                  "items": {}
                },
                "message": {
                  "type": "string",
                  "example": "Authentication failed"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "取得特定使用者",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "使用者 ID"
          }
        ],
        "responses": {
          "200": {
            "description": "成功取得使用者",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "401": {
            "description": "未授權",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Authentication failed!"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "找不到使用者",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "type": "array",
                  "example": [],
                  "items": {}
                },
                "message": {
                  "type": "string",
                  "example": "User not found"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Parents": {
      "type": "object",
      "properties": {
        "father": {
          "type": "string",
          "example": "Simon Doe"
        },
        "mother": {
          "type": "string",
          "example": "Marie Doe"
        }
      }
    },
    "User": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Jhon Doe"
        },
        "age": {
          "type": "number",
          "example": 29
        },
        "parents": {
          "$ref": "#/definitions/Parents"
        },
        "diplomas": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "school": {
                "type": "string",
                "example": "XYZ University"
              },
              "year": {
                "type": "number",
                "example": 2020
              },
              "completed": {
                "type": "boolean",
                "example": true
              },
              "internship": {
                "type": "object",
                "properties": {
                  "hours": {
                    "type": "number",
                    "example": 290
                  },
                  "location": {
                    "type": "string",
                    "example": "XYZ Company"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "security": [
    {
      "apiKeyAuth": []
    }
  ]
}